+++ R/PipeOpEncodeLmer.R
@@ -42,7 +42,8 @@
 #'   otherwise be set during construction. Default `list()`.
 #'
 #' @section Input and Output Channels:
-#' Input and output channels are inherited from [`PipeOpTaskPreproc`].
+#' Input and output channels are inherited from [`PipeOpTaskPreproc`]. Instead of a [`Task`][mlr3::Task], a
+#' [`TaskSupervised`][mlr3::TaskSupervised] is used as input and output during training and prediction.
 #'
 #' The output is the input [`Task`][mlr3::Task] with all affected `factor`, `character` or
 #' `ordered` parameters encoded according to the `method` parameter.
@@ -95,12 +96,14 @@ PipeOpEncodeLmer = R6Class("PipeOpEncodeLmer",
         fast_optim = p_lgl(tags = c("train", "required"))
       )
       ps$values = list(fast_optim = TRUE)
-      super$initialize(id, param_set = ps, param_vals = param_vals, packages = c("lme4", "nloptr"), tags = "encode", feature_types = c("factor", "ordered"))
+      super$initialize(id, param_set = ps, param_vals = param_vals, packages = c("lme4", "nloptr"),
+        task_type = "TaskSupervised", tags = "encode", feature_types = c("factor", "ordered"))
     }
   ),
   private = list(
 
     .get_state_dt = function(dt, levels, target) {
+      # FIXME: Handle non-Regr / non-Classif Tasks that inherit from TaskSupervised, #913
       task_type = if (is.numeric(target)) "regr" else "classif"
       state = list()
       # for prediction, use complete encoding model
@@ -147,6 +150,7 @@ PipeOpEncodeLmer = R6Class("PipeOpEncodeLmer",
       }
       dt_new
     },
+
     fit_lmer = function(feature, target, fast_optim, task_type) {
       args = private$get_args_nlopt_lmer(feature, target, fast_optim, task_type)
       if (task_type == "classif") args$family = stats::binomial
@@ -158,8 +162,8 @@ PipeOpEncodeLmer = R6Class("PipeOpEncodeLmer",
       }
       private$get_coefs(mod)
     },
-    get_args_nlopt_lmer = function(feature, target, fast_optim, task_type) {
 
+    get_args_nlopt_lmer = function(feature, target, fast_optim, task_type) {
       # lmer for regr, glmer for classif
       if (task_type == "regr") {
         control_fun = lme4::lmerControl
@@ -178,6 +182,7 @@ PipeOpEncodeLmer = R6Class("PipeOpEncodeLmer",
         data = data.frame(lvl = feature, y = target),
         na.action = stats::na.omit, control = control)
     },
+
     get_coefs = function(mod) {
       coefs = invoke(stats::coef, mod, .opts = list(warnPartialMatchArgs = FALSE, warnPartialMatchDollar = FALSE))$lvl
       lvls = rownames(coefs)