+++ inst/testthat/helper_functions.R
@@ -217,6 +217,8 @@ expect_pipeop_class = function(poclass, constargs = list(), check_ps_default_val
 #  - training / prediction does not change input task
 #  - is_trained is true
 #  - deep cloning clones the state
+#  - preproc() gives no error with Task indata
+#  - preproc() gives no error with data.table indata if PipeOp works without target column
 #
 # `task` must have at least two feature columns and at least two rows.
 expect_datapreproc_pipeop_class = function(poclass, constargs = list(), task,
@@ -467,6 +469,19 @@ expect_datapreproc_pipeop_class = function(poclass, constargs = list(), task,
 
   if (deterministic_predict && deterministic_train) {
     expect_equal(dtrain, dpredict)
+  }
+
+  # Test that preproc() works with PipeOp
+  expect_no_error(mlr3pipelines::preproc(task, po))
+  expect_no_error(mlr3pipelines::preproc(task, po, predict = TRUE))
+  if (po$input$train == "Task") {  # implies that PipeOp does not require target column
+    expect_no_error({
+      # Need to explicitly refer to mlr3pipelines namespace because this is loaded from inst/
+      dtout = mlr3pipelines::preproc(task$data(cols = task$feature_names), po)
+      expect_data_table(dtout)
+      dtout = mlr3pipelines::preproc(task$data(cols = task$feature_names), po, predict = TRUE)
+      expect_data_table(dtout)
+    })
   }
 }